# PERF: Remove slower short circut checks in RangeIndex__getitem__